+++ webapp/polls/views.py
@@ -1,6 +1,7 @@
 from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.views import generic
+from django.utils import timezone
 from django.shortcuts import render, get_object_or_404
 
 from .models import Question, Choice
@@ -10,13 +11,27 @@ class IndexView(generic.ListView):
     template_name = 'polls/index.html'
     context_object_name = 'questions'
 
+    """
+    Return the last five published questions
+    (not including those set to be published in the future).
+    """
     def get_queryset(self):
-        return Question.objects.order_by('-pub_date')[:5]
+        now = timezone.now()
+        return Question \
+            .objects \
+            .filter(pub_date__lte=now) \
+            .order_by('-pub_date')[:5]
 
 
 class DetailView(generic.DetailView):
     model = Question
     template_name = 'polls/detail.html'
+
+    def get_queryset(self):
+        # Excludes any questions that aren't published yet.
+        return Question \
+            .objects \
+            .filter(pub_date__lte=timezone.now())
 
 
 class ResultsView(generic.DetailView):