+++ HISTORY.md
@@ -40,3 +40,9 @@
 [Writing your first Django app, part 4](https://docs.djangoproject.com/en/4.0/intro/tutorial04/)
 
 1. Added `generic` to `polls/views`
+
+#### Part 5
+
+[Writing your first Django app, part](https://docs.djangoproject.com/en/4.0/intro/tutorial05/)
+
+1. Added Unit Tests